pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "distributed-lakehouse-project"
version = "0.1.0"
description = "Distributed Lakehouse Lab: Spark vs Ray on a 2-core Codespace"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyspark>=3.5.0,<3.6",
    "delta-spark>=3.2.0,<4",
    "ray[default]>=2.20.0",
    "duckdb>=1.0.0",
    "numpy>=1.26.0",
    "pandas>=2.0.0",
    "pyarrow>=15.0.0",
    "psutil>=5.9.0",
    "pyyaml>=6.0",
    "setuptools>=69.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 79
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
